Standardize data on NYPD profile data.
This makes it so the formatting is consistent for name, badge number, command, etc.

This also adds a link to the NYPD profile page for each officer, adds badge number and lists 0 for allegations and substantiations if there's no 50-a match (I double checked the matches and everything looks good).

And this logs an error if someone's profile is missing training data. This happened once, and was just a data error, but this allows the script to keep running.