Match on both NST courses and include tax ID.
This matches on two NST courses, as they did in https://www.nysfocus.com/2022/05/23/eric-adams-nypd-neighborhood-safety-team-roster/ .

This also outputs the tax ID too, so it can be more easily merged with other data that has tax ID (technically the tax ID was already there, in the NYPD Profile link, but this makes it easier to use).

This removes the NST training date since there are two trainings now and instead of trying to figure out a way to include both, I figured it'd be simpler to just remove it since it doesn't seem that important.